requirements.txt
flask==2.3.3
werkzeug==2.3.7
tensorflow==2.12.0
pillow==9.5.0
numpy==1.23.5
python-dotenv==1.0.0